+++ spec/gruf/interceptors/instrumentation/request_logging/formatters/base_spec.rb
@@ -18,10 +18,12 @@ require 'spec_helper'
 
 describe Gruf::Interceptors::Instrumentation::RequestLogging::Formatters::Base do
   let(:formatter) { described_class.new }
+  let(:request) { build :controller_request }
+  let(:result) { Gruf::Interceptors::Timer.time { Rpc::GetThingResponse.new } }
   let(:payload) { {} }
 
   describe '.format' do
-    subject { formatter.format(payload) }
+    subject { formatter.format(payload, request: request, result: result) }
 
     it 'should raise a NotImplementedError' do
       expect { subject }.to raise_error(NotImplementedError)